move ConsentStatus to its own file
